Removed node module, moved functions to workernode
